+++ core-mvi/lib-core-mvi/src/main/java/ru/surfstudio/android/core/mvi/event/lifecycle/LifecycleEventHub.kt
@@ -3,13 +3,14 @@ package ru.surfstudio.android.core.mvi.event.lifecycle
 import android.os.Bundle
 import ru.surfstudio.android.core.mvi.event.Event
 import ru.surfstudio.android.core.mvi.event.hub.EventHub
+import ru.surfstudio.android.core.ui.state.LifecycleStage
 import ru.surfstudio.android.core.ui.state.ScreenState
 
 /**
  * Создатель [LifecycleEvent] на основе [LifecycleStage].
  * Является лямбдой, принимающей [LifecycleStage] и создающей на его основе событие.
  */
-typealias LifecycleEventCreator<T> = (LifecycleStage) -> T
+typealias LifecycleEventCreator<E> = (LifecycleStage) -> E
 
 /**
  * [EventHub], рассылюащий события жизненного цикла экрана, к которому прикреплен
@@ -22,39 +23,39 @@ interface LifecycleEventHub<T : Event, Stream> : EventHub<T, Stream>, LifecycleV
 
     override fun onViewReady() {
         if (!screenState.isViewRecreated) {
-            emitEventByType(LifecycleStage.CREATE)
+            emitEventByType(LifecycleStage.CREATED)
         }
-        emitEventByType(LifecycleStage.VIEW_CREATE)
+        emitEventByType(LifecycleStage.VIEW_CREATED)
     }
 
     override fun onStart() {
-        emitEventByType(LifecycleStage.START)
+        emitEventByType(LifecycleStage.STARTED)
     }
 
     override fun onResume() {
-        emitEventByType(LifecycleStage.RESUME)
+        emitEventByType(LifecycleStage.RESUMED)
     }
 
     override fun onPause() {
-        emitEventByType(LifecycleStage.PAUSE)
+        emitEventByType(LifecycleStage.PAUSED)
     }
 
     override fun onStop() {
-        emitEventByType(LifecycleStage.STOP)
+        emitEventByType(LifecycleStage.STOPPED)
     }
 
     override fun onViewDestroy() {
-        emitEventByType(LifecycleStage.VIEW_DESTROY)
+        emitEventByType(LifecycleStage.VIEW_DESTROYED)
     }
 
     override fun onCompletelyDestroy() {
-        emitEventByType(LifecycleStage.DESTROY)
+        emitEventByType(LifecycleStage.COMPLETELY_DESTROYED)
     }
 
     fun emitEventByType(lifecycleStage: LifecycleStage) {
-        when (val event = lifecycleEventCreator?.invoke(lifecycleStage)) {
-            null -> return
-            else -> emit(event)
+        val event = lifecycleEventCreator?.invoke(lifecycleStage)
+        if (event != null) {
+            emit(event)
         }
     }