+++ core-mvi/sample/src/main/java/ru/surfstudio/android/core/mvi/sample/ui/base/middleware/experimental/navigation/close/CloseAffinityEvent.kt
@@ -1,0 +1,9 @@
+package ru.surfstudio.android.core.mvi.sample.ui.base.middleware.experimental.navigation.close
+
+import ru.surfstudio.android.core.mvi.sample.ui.base.middleware.experimental.ExperimentalFeature
+
+/**
+ * Событие закрытия таска экранов
+ */
+@ExperimentalFeature
+interface CloseAffinityEvent : CloseScreenEvent