+++ core-mvi/RELEASE_NOTES.md
@@ -2,4 +2,5 @@
 # Core-mvi Release Notes
 ## 0.5.0-alpha
 ##### Core-mvi
-* Core MVI Refactor was done
+* ANDDEP-671 Core mvi refactor, add comments
+* ANDDEP-671 Add navigation middleware, add dsl